Revert "Revert "ref(control_silo): move logic from old to new slack tasks (#75549)""
This reverts commit 835590895c8b63337716f40e30c5f3b801ded3c1.
