+++ pyproject.toml
@@ -373,8 +373,6 @@ module = [
     "sentry.tasks.auth",
     "sentry.tasks.base",
     "sentry.tasks.integrations",
-    "sentry.tasks.integrations.slack.find_channel_id_for_rule",
-    "sentry.tasks.integrations.slack.link_slack_user_identities",
     "sentry.tasks.process_buffer",
     "sentry.tasks.sentry_apps",
     "sentry.templatetags.sentry_assets",
