+++ src/sentry/api/endpoints/project_rule_details.py
@@ -28,6 +28,7 @@ from sentry.apidocs.parameters import GlobalParams, IssueAlertParams
 from sentry.constants import ObjectStatus
 from sentry.integrations.jira.actions.create_ticket import JiraCreateTicketAction
 from sentry.integrations.jira_server.actions.create_ticket import JiraServerCreateTicketAction
+from sentry.integrations.slack.tasks.find_channel_id_for_rule import find_channel_id_for_rule
 from sentry.integrations.slack.utils import RedisRuleStatus
 from sentry.mediators.project_rules.updater import Updater
 from sentry.models.rule import NeglectedRule, RuleActivity, RuleActivityType
@@ -35,7 +36,6 @@ from sentry.models.scheduledeletion import RegionScheduledDeletion
 from sentry.rules.actions import trigger_sentry_app_action_creators_for_issues
 from sentry.rules.actions.utils import get_changed_data, get_updated_rule_data
 from sentry.signals import alert_rule_edited
-from sentry.tasks.integrations.slack import find_channel_id_for_rule
 from sentry.types.actor import Actor
 from sentry.utils import metrics
