+++ src/sentry/incidents/endpoints/organization_alert_rule_details.py
@@ -30,10 +30,12 @@ from sentry.incidents.logic import (
 )
 from sentry.incidents.serializers import AlertRuleSerializer as DrfAlertRuleSerializer
 from sentry.incidents.utils.sentry_apps import trigger_sentry_app_action_creators_for_incidents
+from sentry.integrations.slack.tasks.find_channel_id_for_alert_rule import (
+    find_channel_id_for_alert_rule,
+)
 from sentry.integrations.slack.utils import RedisRuleStatus
 from sentry.models.rulesnooze import RuleSnooze
 from sentry.sentry_apps.services.app import app_service
-from sentry.tasks.integrations.slack import find_channel_id_for_alert_rule
 from sentry.users.services.user.service import user_service
